requirements.txt
aiogram==2.25.1
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.6.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
Babel==2.9.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Django==5.0.1
django-cors-headers==4.3.1
django-jazzmin==2.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dotty-dict==1.3.1
drf-spectacular==0.27.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.41
idna==3.6
importlib-resources==6.1.1
inflection==0.5.1
Jinja2==3.1.3
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
magic-filter==1.0.12
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
multidict==6.0.4
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
PyJWT==2.8.0
python-dotenv==1.0.1
python-gitlab==4.4.0
python-semantic-release==8.7.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.0
rpds-py==0.17.1
shellingham==1.5.4
smmap==5.0.1
sqlparse==0.4.4
tomlkit==0.12.3
twilio==8.11.1
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
yarl==1.9.4
